fix getConfirmBookingPageByMember, convertToTicketByMember, getConfirmBookingPageByMember

+++ MovieBooking 2/src/main/java/com/example/MovieBooking/entity/dto/BookingDTO.java
@@ -24,7 +24,7 @@ public class BookingDTO {
     private int status;
     private String screen;
     private long total;
-    private int useScore;
+    private Long useScore;
 
 
     @Override
